Add basic info viewing to dashboard
Adds github service to hit api
and retrieve data for required
queries. Adds both a repository
and a follow PORO to to interact
with github_service and return
appropriately formatted data.
Methods added to user which
interacted with necessary POROs
to complete calls and return
data for the appropriate user.